+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added zoom methods to gl-map.
 - Made gl-form-page and gl-draw-toolbar responsive.
 - Wrap text in form elements.
+- Trigger an event on gl-like-button click.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.